requirements.txt
PySide6>=6.6.0
easyocr>=1.7.1
torch
torchvision
numpy<2.0
pillow>=9.0
mss>=9.0
pywinctl>=0.4
